+++ components/__tests__/jschema/properties/anyOf.test.js
@@ -1,9 +1,9 @@
 import { describe, it, expect } from 'vitest';
 import { renderSchema } from './property_test_utils';
-import { screen } from '@testing-library/svelte';
+import { fireEvent, screen } from '@testing-library/svelte';
 
 describe('anyOf properties', () => {
-	it('Nullable property should use parent title and description', async () => {
+	it('Nullable object', async () => {
 		const { component } = renderSchema(
 			{
 				$defs: {
@@ -33,9 +33,16 @@ describe('anyOf properties', () => {
 			nullable_arg_with_null_default: null
 		});
 
+		// Should use parent title and description
 		expect(screen.getByText('Nullable Arg With Null Default')).toBeVisible();
 		expect(screen.getByLabelText('Description').getAttribute('data-bs-content')).eq(
 			'Type hint SimpleModel | None = None'
 		);
+
+		await fireEvent.click(screen.getByRole('switch', { name: 'Set' }));
+
+		expect(component.getArguments()).deep.eq({
+			nullable_arg_with_null_default: { foo: null }
+		});
 	});
 });