+++ components/src/lib/jschema/form_element.js
@@ -139,6 +139,15 @@ export class UnexpectedFormElement extends ValueFormElement {
 	}
 }
 
+export class UnsupportedFormElement extends ValueFormElement {
+	/**
+	 * @param {import("../types/form").ValueFormElementFields<any>} fields
+	 */
+	constructor(fields) {
+		super(fields);
+	}
+}
+
 export class InvalidFormElement extends ValueFormElement {
 	/**
 	 * @param {import("../types/form").ValueFormElementFields<any>} fields
@@ -182,7 +191,7 @@ export class ObjectFormElement extends NullableFormElement {
 		 */
 		this.additionalProperties = fields.additionalProperties;
 		this.children = fields.children;
-		this.collapsed = !fields.required;
+		this.collapsed = !fields.required || fields.isNull;
 	}
 
 	/**
@@ -346,7 +355,7 @@ export class ArrayFormElement extends NullableFormElement {
 		this.items = fields.items;
 		this.minItems = fields.minItems;
 		this.maxItems = fields.maxItems;
-		this.collapsed = !fields.required;
+		this.collapsed = !fields.required || fields.isNull;
 	}
 
 	addChild() {
@@ -435,7 +444,7 @@ export class TupleFormElement extends NullableFormElement {
 		super(fields);
 		this.children = fields.children;
 		this.items = fields.items;
-		this.collapsed = !fields.required;
+		this.collapsed = !fields.required || fields.isNull;
 	}
 
 	addTuple() {